requirements.txt
click==8.1.7
h5py==3.11.0
lightning==2.3.3
natsort==8.4.0
numpy==2.1.0
pandas==2.2.2
pytorch_lightning==2.3.3
PyYAML==6.0.2
scikit_learn==1.5.1
torch==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
wandb==0.17.7